refactor(routes): consolidate frontend route definitions - move auth, tenant, and owner route declarations into AppRouter - inline login redirect handling and remove route helper modules - update protected role redirects and post-auth destinations - add router tests for owner redirection and removed fallback routes -create the new owner pages
