--- frontend/src/pages/owner/OwnerComingSoonPage.tsx
@@ -1,22 +1,0 @@
-import logo from '@/assets/logo.png'
-
-export default function OwnerComingSoonPage() {
-  return (
-    <main className="relative flex min-h-screen items-center justify-center bg-[var(--tenant-bg)] p-4 sm:p-8">
-      <div className="w-full max-w-3xl overflow-hidden rounded-3xl border border-[var(--tenant-border-light)] bg-white/90 shadow-2xl shadow-slate-950/5 backdrop-blur-sm">
-        <section className="grid min-h-[420px] items-center gap-8 p-8 sm:p-12">
-          <div className="text-center">
-            <img src={logo} alt="Roomies" className="mx-auto h-14 w-auto" />
-            <h1 className="mt-6 text-3xl font-bold tracking-tight text-[var(--rm-text-strong)] sm:text-4xl">
-              Perfil de propietario proximamente
-            </h1>
-            <p className="mt-4 text-sm leading-relaxed text-[var(--rm-text-soft)] sm:text-base">
-              Estamos construyendo el onboarding para propietarios. Puedes seguir explorando la aplicacion mientras
-              terminamos esta seccion.
-            </p>
-          </div>
-        </section>
-      </div>
-    </main>
-  )
-}
