+++ frontend/src/routes/paths.ts
@@ -11,9 +11,12 @@ export const paths = {
   tenantMessages: '/tenant/messages',
   tenantNotifications: '/tenant/notifications',
   tenantProfile: '/tenant/profile',
-  ownerComingSoon: '/owner/coming-soon',
-  ownerDashboard: '/owner/dashboard',
+  ownerProperties: '/owner/properties',
   ownerPublishProperty: '/owner/properties/new',
+  ownerApplications: '/owner/applications',
+  ownerMessages: '/owner/messages',
+  ownerNotifications: '/owner/notifications',
+  ownerProfile: '/owner/profile',
 } as const
 
 export type AppPath = (typeof paths)[keyof typeof paths]
